backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]>=0.34.0
sse-starlette==1.8.2
pydantic==2.9.2
google-cloud-storage==2.18.2
google-cloud-firestore==2.17.0
google-cloud-pubsub==2.21.5
google-cloud-aiplatform>=1.87.0
google-auth>=2.35.0
python-multipart==0.0.9
google-adk==0.2.0

